изменид вьюмодели

+++ app/src/main/java/com/confectinary/app/db/entity/relation/manager_with_confectionary/ConfectionaryAndManagerCrossRef.kt
@@ -9,9 +9,9 @@ import androidx.room.Entity
 )
 data class ConfectionaryAndManagerCrossRef(
     @ColumnInfo(name = "manager_id")
-    val managerId: Int,
+    val managerId: Long,
 
     @ColumnInfo(name = "confectionary_id")
-    val confectionaryId: Int,
+    val confectionaryId: Long,
 
 )
