+++ app/src/main/java/com/confectinary/app/db/entity/relation/pastry_with_confectioner/PastryAndConfectionerCrossRef.kt
@@ -9,7 +9,7 @@ import androidx.room.Entity
 )
 data class PastryAndConfectionerCrossRef(
     @ColumnInfo(name = "confectioner_id")
-    val confectionerId: Int,
+    val confectionerId: Long,
     @ColumnInfo(name = "pastry_id")
-    val pastryId: Int
+    val pastryId: Long
 )
